+++ gioco.py
@@ -1,5 +1,5 @@
 # gioco.py
-# Interprete Interattivo per FAVELLA 1 (v0.33.0)
+# Interprete Interattivo per FAVELLA 1 (v0.34.0)
 
 import sys
 import traceback